PDF_Summarizer/requirements.txt
openai
langchain
langchain-openai
langchain_community
PyPDF2
faiss-cpu
tiktoken
streamlit
python-dotenv